requirements.txt
pymongo
fastapi
requests
uvicorn
python-dotenv
python-multipart
